fix: blocked redirect to login
This commit fixes the following issue. When a session is
invalidated/expires, a refresh results in a server error instead of
redirecting the user appropriately to the login page.

The true source of the 'bug', if one can even call it that, is uncertain
(although this analysis should not be too far off the mark). Some
investigation indicates that the effect that is supposed to handle an
invalidated session (in `(authed)/+layout.svelte`) does not mount before
a `+page.ts` call to `apiClient` to fetch data is made. So while the
fetch is made and data is returned (an empty body is returned with a 401
Unauthorized code), the data is processed, causing a crash as no error
handling has been implemented yet (separate issue). This crash prevents
the effect from ever mounting.

Disabling SSR appears to fix this --- it forces the effect to mount.

Either way, disabling SSR is sensible for this entire route group since
data loaded is user-dependent.